+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 * Event filter and GroundTruth modules: add brief documentation.
 * Registration matrix viewer and manual editor: adjust text for control modifier to the OS (CTRL for Linux/Windows and CMD for Mac OS X) in the documentation.
 * Registration module: More explicit error message when registration shift is too large resulting in an empty registered image after cropping.
-
+* Metadata viewer: fix file type filter in file selection dialog.
 
 
 ## [v2.0.0] 2024-07-12